vastai/requirements.txt
# vastai setup
paramiko
GitPython
vastai_sdk
fabric